Backport of fcd050d4abc3a43e04aa7f0382abdfe128615d21
